fixed location and add react 'origin/master'

+++ app/src/main/java/com/example/greendefend/domin/model/forum/React.kt
@@ -7,5 +7,5 @@ data class React(
     @SerializedName("userID")
     val userId: String,
     @SerializedName("postID")
-    val postId: Long,
+    val postId: Int,
 )
